Ansible flag variables for physical and virtual machines
In some cases, Ansible was misidentifying the virtualization type of DO VM's as
'host', leading to extra steps being run in playbooks unnecessarily. Also, the
code to check for being in a virtual machine was becoming repetitive.

By extracting the flags to variables, both of these issues are side stepped.